// src/app/Module/cliente/components/auth/models/persona-request.model.ts
export class PersonaRequest {
    IdPersona: number= 0;
    IdPersonTipoDocumento?: number=0;
    NroDocumento?: string ="";
    Nombre?: string ="";
    ApPaterno?: string ="";
    ApMaterno?: string ="";
    NombreCompleto?: string ="";
    FechaNacimiento?: string =""; 
    IdGenero?: number=0;
    Email?: string ="";
    Celular?: string ="";
    UsuarioCrea?: string ="";
    UsuarioActualiza?: string ="";
    FechaCrea?: string =""; 
    FechaActualiza?: string =""; 
    IdPersonTipo?: number=0;
    IdUbica?: number=0;
}